contrib/cluster_manager/pyproject.toml
[project]
name = "cluster-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "paramiko>=4.0.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
cluster-manager = "main:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
